third_party/blink/web_tests/ppapi/plugins/plugin-document-back-forward-expected.txt

CONSOLE MESSAGE: Blink Test Plugin: initializing
CONSOLE MESSAGE: Blink Test Plugin: initializing
PASS successfullyParsed is true

TEST COMPLETE
